+++ services/web/tmtar/events/constants.py
@@ -1,0 +1,5 @@
+class Constants(object):
+    event_type_length = 10
+    title_max_length = 200
+    title_min_length = 10
+    short_desc_max_length = 500