+++ services/web/tmtar/events/interface_test.py
@@ -18,7 +18,7 @@ def interface() -> IEvent:
         event_type=EventType[0],
         location_id=1,
         update_date=datetime.utcnow(),
-        title="Sample",
+        title="Sample_event_title",
         short_description="""The plugin adds a random text generator, capable 
                  of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                  or replace a selection.""",