+++ services/web/tmtar/events/model_test.py
@@ -17,7 +17,7 @@ def event():
         event_type=EventType[0],
         location_id=1,
         update_date=datetime.utcnow(),
-        title="Sample",
+        title="Sample_event_title",
         short_description="""The plugin adds a random text generator, capable 
                  of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                  or replace a selection.""",