+++ services/web/tmtar/events/schema_test.py
@@ -25,7 +25,7 @@ def test_schema_works(schema: EventSchema):
         dict(
             event_type=EventType[0],
             location_id='1',
-            title="Sample",
+            title="Sample_event_title",
             short_description="""The plugin adds a random text generator, capable 
                      of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                      or replace a selection.""",
@@ -41,7 +41,7 @@ def test_schema_works(schema: EventSchema):
 
     assert event.event_type == EventType[0]
     assert event.location_id == 1
-    assert event.title == "Sample"
+    assert event.title == "Sample_event_title"
     assert event.short_description == """The plugin adds a random text generator, capable 
                      of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                      or replace a selection."""
@@ -59,7 +59,7 @@ def test_update_schema_works():
         dict(
             event_type=EventType[0],
             location_id='1',
-            title="Sample",
+            title="Sample_event_title",
             short_description="""The plugin adds a random text generator, capable 
                      of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                      or replace a selection.""",
@@ -76,7 +76,7 @@ def test_update_schema_works():
 
     assert event.event_type == EventType[0]
     assert event.location_id == 1
-    assert event.title == "Sample"
+    assert event.title == "Sample_event_title"
     assert event.short_description == """The plugin adds a random text generator, capable 
                      of creating witty texts in different genres. Created text can be inserted newly at the caret, 
                      or replace a selection."""