+++ services/web/tmtar/files/schema.py
@@ -3,4 +3,4 @@ from marshmallow import Schema, fields
 
 class FileSchema(Schema):
     id = fields.Number(attribute='id', dump_only=True)
-    filename = fields.String(attribute='filename')
+    filename = fields.String(attribute='filename', dump_only=True)