+++ services/web/tmtar/files/schema_test.py
@@ -11,10 +11,3 @@ def schema() -> FileSchema:
 
 def test_FileSchema_create(schema: FileSchema):  # noqa
     assert schema
-
-
-def test_FileSchema_works(schema: FileSchema):  # noqa
-    params = schema.load({'filename': 'testfile.txt'})
-    widget = File(**params)
-
-    assert widget.filename == 'testfile.txt'