+++ services/web/tmtar/zones/constants.py
@@ -1,0 +1,11 @@
+class Constants(object):
+    title_max_length = 200
+    title_min_length = 5
+    description_max_length = 500
+    SRID = 4326  # noqa: WPS115
+    latitude_field = 'latitude'
+    longitude_field = 'longitude'
+    latitude_limit = 90
+    longitude_limit = 180
+    max_radius = 2000
+    min_radius = 20