+++ services/web/tmtar/zones/schema.py
@@ -1,48 +1,53 @@
 from geoalchemy2 import func, shape
 from marshmallow import Schema, fields, post_load, pre_dump, validate
 
+from .constants import Constants
 from .model import Zone
-
-latitude_field = 'latitude'
-longitude_field = 'longitude'
-latitude_limit = 90
-longitude_limit = 180
-max_radius = 2000
 
 
 class ZoneSchema(Schema):
     """Zone instance schema."""
 
     id = fields.Integer(attribute='id', dump_only=True)
-    title = fields.String(attribute='title', required=True)
+    title = fields.String(
+        attribute='title',
+        required=True,
+        validate=validate.Length(Constants.title_min_length, Constants.title_max_length),
+    )
     location_id = fields.Integer(attribute='location_id', required=True)
     latitude = fields.Float(
-        attribute=latitude_field,
-        validate=validate.Range(-latitude_limit, latitude_limit),
+        attribute=Constants.latitude_field,
+        validate=validate.Range(-Constants.latitude_limit, Constants.latitude_limit),
         required=True,
     )
     longitude = fields.Float(
-        attribute=longitude_field,
-        validate=validate.Range(-longitude_limit, longitude_limit),
+        attribute=Constants.longitude_field,
+        validate=validate.Range(-Constants.longitude_limit, Constants.longitude_limit),
         required=True,
     )
     radius = fields.Float(
         attribute='radius',
         required=True,
-        validate=validate.Range(0, max_radius),
+        validate=validate.Range(Constants.min_radius, Constants.max_radius),
     )
     active = fields.Boolean(attribute='active')
     secret = fields.Boolean(attribute='secret')
+    description = fields.String(
+        attribute='description',
+        validate=validate.Length(max=Constants.description_max_length),
+    )
+    actual_address = fields.String(attribute='actual_address')
+    preview_image_filename = fields.String(attribute='preview_image_filename')
 
     @post_load
     def make_point(self, load_data: dict, **kwargs):
         """Make a point from latitude and longitude."""
         load_data['center'] = func.ST_MakePoint(
-            load_data[longitude_field],
-            load_data[latitude_field],
+            load_data[Constants.longitude_field],
+            load_data[Constants.latitude_field],
         )
-        load_data.pop(latitude_field)
-        load_data.pop(longitude_field)
+        load_data.pop(Constants.latitude_field)
+        load_data.pop(Constants.longitude_field)
         return load_data
 
     @pre_dump
@@ -60,31 +65,34 @@ class ZoneUpdateSchema(ZoneSchema):
     title = fields.String(attribute='title')
     location_id = fields.Integer(attribute='location_id')
     latitude = fields.Float(
-        attribute=latitude_field,
-        validate=validate.Range(-latitude_limit, latitude_limit),
+        attribute=Constants.latitude_field,
+        validate=validate.Range(-Constants.latitude_limit, Constants.latitude_limit),
     )
     longitude = fields.Float(
-        attribute=longitude_field,
-        validate=validate.Range(-longitude_limit, longitude_limit),
+        attribute=Constants.longitude_field,
+        validate=validate.Range(-Constants.longitude_limit, Constants.longitude_limit),
     )
     radius = fields.Float(
         attribute='radius',
-        validate=validate.Range(0, max_radius),
+        validate=validate.Range(Constants.min_radius, Constants.max_radius),
     )
 
     @post_load
     def make_point(self, load_data: dict, **kwargs):
         """Make a point from latitude and longitude."""
-        if longitude_field in load_data.keys() and latitude_field in load_data.keys():
+        latitude_presence = Constants.latitude_field in load_data.keys()
+        longitude_presence = Constants.longitude_field in load_data.keys()
+
+        if latitude_presence and longitude_presence:
             load_data['center'] = func.ST_MakePoint(
-                load_data[longitude_field],
-                load_data[latitude_field],
+                load_data[Constants.longitude_field],
+                load_data[Constants.latitude_field],
             )
 
-        if longitude_field in load_data.keys():
-            load_data.pop(longitude_field)
+        if longitude_presence:
+            load_data.pop(Constants.longitude_field)
 
-        if latitude_field in load_data.keys():
-            load_data.pop(latitude_field)
+        if latitude_presence:
+            load_data.pop(Constants.latitude_field)
 
         return load_data